Refactored enums
